# api/swagger.yaml
openapi: 3.0.2
info:
  title: Dancier
  version: '1.1'
servers:
  - url: 'https://dancier.net'
  - url: 'http://localhost:8080'
paths:
...
  /dancers:
    get:
      summary: retrieves all dancers in a predefined range
      tags:
        - Dancers
      parameters:
        - $ref: '#/components/parameters/range'
        - $ref: '#/components/parameters/gender'
      responses:
        '200':
          description: retrieves a potentially empty array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/public_profile"
    post:
      summary: Used to retrieve a list of dancers
      tags:
        - Dancers
        - Chats
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dancerListRequest'
      responses:
        '201':
          description: "list of dancers created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dancerListResponse'
              example:
                123:
                  id: 123
                  dancerName: dancero
                  city: Dortmund
                  profileImageHash: 10887E4B-293E-4837-8132-CA85C942316F
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  securitySchemes:
    jwt-token:
      type: apiKey
      in: cookie
      name: jwt-token
...
  parameters:
...
    gender:
      name: gender
      in: query
      description: required preferred gender of the dancers
      schema:
        type: string
    range:
      name: range
      in: query
      description: Range [km] in which the search for dancers should be done, defaults to 20 km
      schema:
        type: string
  schemas:
...
    gender:
      type: string
      enum:
        - MALE
        - FEMALE
        - DIVERS
        - NA
...
    dance_profile:
      type: object
      properties:
        dance:
          type: string
        level:
          type: string
          enum:
            - NO_EXPERIENCE
            - BASIC
            - INTERMEDIATE
            - ADVANCED
            - PRO
        leading:
          type: string
          enum:
            - FOLLOW
            - LEAD
            - BOTH
            - NA
...
    public_profile:
      type: object
      properties:
        id:
          type: string
        aboutMe:
          type: string
        size:
          type: integer
        gender:
          $ref: "#/components/schemas/gender"
        dancerName:
          type: string
        age:
          type: string
        ableTo:
          type: array
          items:
            $ref: '#/components/schemas/dance_profile'
        wantsTo:
          type: array
          items:
            $ref: '#/components/schemas/dance_profile'
        city:
          type: string
        country:
          type: string
          enum:
            - GER
        profileImageHash:
          type: string
...
    dancerListRequest:
      type: object
      properties:
        dancer_ids:
          type: array
          items:
            type: string
            format: uuid
    dancerListResponse:
      type: object
      additionalProperties:
        type: object
        properties:
          id:
            type: string
            format: uuid
          dancerName:
            type: string
          city:
            type: string
          profileImageHash:
            type: string
...
    error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message